+++ Expanded Cheat Menu/Changelog.md
@@ -16,6 +16,7 @@
 - Rocket>Change Resupply Settings (thanks XxUnkn0wnxX).
 - Oxygen Need/Free wasn't working (thanks XxUnkn0wnxX).
 - Shuttle speed broke in DA (thanks hchsiao).
+- Shuttle Capacity broke in DA (thanks hchsiao).
 
 ## v8.5 (26 Aug 2018)
 ### Added: